+++ R/chat.R
@@ -4,7 +4,7 @@ NULL
 #' A chat
 #'
 #' @description
-#' A `Chat` is an sequence of sequence of user and assistant [Turn]s sent
+#' A `Chat` is a sequence of user and assistant [Turn]s sent
 #' to a specific [Provider]. A `Chat` is a mutable R6 object that takes care of
 #' managing the state associated with the chat; i.e. it records the messages
 #' that you send to the server, and the messages that you receive back.
@@ -18,7 +18,8 @@ NULL
 #' @examplesIf has_credentials("openai")
 #' chat <- chat_openai(echo = TRUE)
 #' chat$chat("Tell me a funny joke")
-Chat <- R6::R6Class("Chat",
+Chat <- R6::R6Class(
+  "Chat",
   public = list(
     #' @param provider A provider object.
     #' @param turns An unnamed list of turns to start the chat with (i.e.,
@@ -108,33 +109,34 @@ Chat <- R6::R6Class("Chat",
       invisible(self)
     },
 
-    #' @description A data frame with a `tokens` column that proides the 
-    #'   number of input tokens used by user turns and the number of 
+    #' @description A data frame with a `tokens` column that proides the
+    #'   number of input tokens used by user turns and the number of
     #'   output tokens used by assistant turns.
-    #' @param include_system_prompt Whether to include the system prompt in 
+    #' @param include_system_prompt Whether to include the system prompt in
     #'   the turns (if any exists).
     tokens = function(include_system_prompt = FALSE) {
       turns <- self$get_turns(include_system_prompt = FALSE)
       assistant_turns <- keep(turns, function(x) x@role == "assistant")
-      if (length(assistant_turns) == 0) {
-        return(data.frame(role = character(), tokens = double()))
-      }
 
       n <- length(assistant_turns)
-      tokens <- t(vapply(assistant_turns, function(turn) turn@tokens, double(2)))
+      tokens <- t(vapply(
+        assistant_turns,
+        function(turn) turn@tokens,
+        double(2)
+      ))
       if (n > 1) {
         # Compute just the new tokens
-        tokens[-1, 1] <- tokens[seq(2, n), 1] - 
+        tokens[-1, 1] <- tokens[seq(2, n), 1] -
           (tokens[seq(1, n - 1), 1] + tokens[seq(1, n - 1), 2])
       }
       # collapse into a single vector
       tokens_v <- c(t(tokens))
-      
+
       tokens_df <- data.frame(
         role = rep(c("user", "assistant"), times = n),
         tokens = tokens_v
       )
-      
+
       if (include_system_prompt && private$has_system_prompt()) {
         # How do we compute this?
         tokens_df <- rbind(data.frame(role = "system", tokens = 0), tokens_df)
@@ -152,7 +154,8 @@ Chat <- R6::R6Class("Chat",
       role <- arg_match(role)
 
       n <- length(private$.turns)
-      switch(role,
+      switch(
+        role,
         system = if (private$has_system_prompt()) private$.turns[[1]],
         assistant = if (n > 1) private$.turns[[n]],
         user = if (n > 1) private$.turns[[n - 1]]
@@ -172,7 +175,11 @@ Chat <- R6::R6Class("Chat",
 
       # Returns a single turn (the final response from the assistant), even if
       # multiple rounds of back and forth happened.
-      coro::collect(private$chat_impl(turn, stream = echo != "none", echo = echo))
+      coro::collect(private$chat_impl(
+        turn,
+        stream = echo != "none",
+        echo = echo
+      ))
 
       text <- self$last_turn()@text
       if (echo == "none") text else invisible(text)
@@ -280,7 +287,12 @@ Chat <- R6::R6Class("Chat",
 
       map(json, function(json) {
         turn <- value_turn(private$provider, json, has_type = TRUE)
-        extract_data(turn, type, convert = convert, needs_wrapper = needs_wrapper)
+        extract_data(
+          turn,
+          type,
+          convert = convert,
+          needs_wrapper = needs_wrapper
+        )
       })
     },
 
@@ -355,7 +367,7 @@ Chat <- R6::R6Class("Chat",
     },
 
     #' @description Register a tool (an R function) that the chatbot can use.
-    #'   If the chatbot decides to use the function,  ellmer will automatically
+    #'   If the chatbot decides to use the function, ellmer will automatically
     #'   call it and submit the results back.
     #'
     #'   The return value of the function. Generally, this should either be a
@@ -370,6 +382,38 @@ Chat <- R6::R6Class("Chat",
       }
 
       private$tools[[tool_def@name]] <- tool_def
+      invisible(self)
+    },
+
+    #' @description Get the underlying provider object. For expert use only.
+    get_provider = function() {
+      private$provider
+    },
+
+    #' @description Retrieve the list of registered tools.
+    get_tools = function() {
+      private$tools
+    },
+
+    #' @description Sets the available tools. For expert use only; most users
+    #'   should use `register_tool()`.
+    #'
+    #' @param tools A list of tool definitions created with [ellmer::tool()].
+    set_tools = function(tools) {
+      if (!is_list(tools) || !all(map_lgl(tools, S7_inherits, ToolDef))) {
+        msg <- "{.arg tools} must be a list of tools created with {.fn ellmer::tool}."
+        if (S7_inherits(tools, ToolDef)) {
+          msg <- c(msg, "i" = "Did you mean to call {.code $register_tool()}?")
+        }
+        cli::cli_abort(msg)
+      }
+
+      private$tools <- list()
+
+      for (tool_def in tools) {
+        self$register_tool(tool_def)
+      }
+
       invisible(self)
     }
   ),
@@ -391,16 +435,29 @@ Chat <- R6::R6Class("Chat",
       if (private$.turns[[i]]@role != "user") {
         private$.turns[[i + 1]] <- Turn("user", contents)
       } else {
-        private$.turns[[i]]@contents <- c(private$.turns[[i]]@contents, contents)
+        private$.turns[[i]]@contents <- c(
+          private$.turns[[i]]@contents,
+          contents
+        )
       }
       invisible(self)
     },
 
     # If stream = TRUE, yields completion deltas. If stream = FALSE, yields
     # complete assistant turns.
-    chat_impl = generator_method(function(self, private, user_turn, stream, echo) {
-      while(!is.null(user_turn)) {
-        for (chunk in private$submit_turns(user_turn, stream = stream, echo = echo)) {
+    chat_impl = generator_method(function(
+      self,
+      private,
+      user_turn,
+      stream,
+      echo
+    ) {
+      while (!is.null(user_turn)) {
+        for (chunk in private$submit_turns(
+          user_turn,
+          stream = stream,
+          echo = echo
+        )) {
           yield(chunk)
         }
         user_turn <- private$invoke_tools()
@@ -409,9 +466,19 @@ Chat <- R6::R6Class("Chat",
 
     # If stream = TRUE, yields completion deltas. If stream = FALSE, yields
     # complete assistant turns.
-    chat_impl_async = async_generator_method(function(self, private, user_turn, stream, echo) {
-      while(!is.null(user_turn)) {
-        for (chunk in await_each(private$submit_turns_async(user_turn, stream = stream, echo = echo))) {
+    chat_impl_async = async_generator_method(function(
+      self,
+      private,
+      user_turn,
+      stream,
+      echo
+    ) {
+      while (!is.null(user_turn)) {
+        for (chunk in await_each(private$submit_turns_async(
+          user_turn,
+          stream = stream,
+          echo = echo
+        ))) {
           yield(chunk)
         }
         user_turn <- await(private$invoke_tools_async())
@@ -423,8 +490,14 @@ Chat <- R6::R6Class("Chat",
 
     # If stream = TRUE, yields completion deltas. If stream = FALSE, yields
     # complete assistant turns.
-    submit_turns = generator_method(function(self, private, user_turn, stream, echo, type = NULL) {
-
+    submit_turns = generator_method(function(
+      self,
+      private,
+      user_turn,
+      stream,
+      echo,
+      type = NULL
+    ) {
       if (echo == "all") {
         cat_line(format(user_turn), prefix = "> ")
       }
@@ -466,7 +539,11 @@ Chat <- R6::R6Class("Chat",
           cat_line(formatted, prefix = "< ")
         }
       } else {
-        turn <- value_turn(private$provider, response, has_type = !is.null(type))
+        turn <- value_turn(
+          private$provider,
+          response,
+          has_type = !is.null(type)
+        )
         text <- turn@text
         if (!is.null(text)) {
           text <- paste0(text, "\n")
@@ -484,7 +561,14 @@ Chat <- R6::R6Class("Chat",
 
     # If stream = TRUE, yields completion deltas. If stream = FALSE, yields
     # complete assistant turns.
-    submit_turns_async = async_generator_method(function(self, private, user_turn, stream, echo, type = NULL) {
+    submit_turns_async = async_generator_method(function(
+      self,
+      private,
+      user_turn,
+      stream,
+      echo,
+      type = NULL
+    ) {
       response <- chat_perform(
         provider = private$provider,
         mode = if (stream) "async-stream" else "async-value",
@@ -558,29 +642,36 @@ is_chat <- function(x) {
 #' @export
 print.Chat <- function(x, ...) {
   turns <- x$get_turns(include_system_prompt = TRUE)
-  
+
   tokens <- x$tokens(include_system_prompt = TRUE)
   tokens_user <- sum(tokens$tokens[tokens$role == "user"])
   tokens_assistant <- sum(tokens$tokens[tokens$role == "assistant"])
 
   cat(paste0(
-    "<Chat", 
-    " turns=", length(turns), 
-    " tokens=", tokens_user, "/", tokens_assistant, 
+    "<Chat",
+    " turns=",
+    length(turns),
+    " tokens=",
+    tokens_user,
+    "/",
+    tokens_assistant,
     ">\n"
   ))
-  
+
   for (i in seq_along(turns)) {
     turn <- turns[[i]]
-    
-    color <- switch(turn@role,
+
+    color <- switch(
+      turn@role,
       user = cli::col_blue,
       assistant = cli::col_green,
       system = cli::col_br_white,
       identity
     )
 
-    cli::cat_rule(cli::format_inline("{color(turn@role)} [{tokens$tokens[[i]]}]"))
+    cli::cat_rule(cli::format_inline(
+      "{color(turn@role)} [{tokens$tokens[[i]]}]"
+    ))
     for (content in turn@contents) {
       cat_line(format(content))
     }
@@ -589,7 +680,10 @@ print.Chat <- function(x, ...) {
   invisible(x)
 }
 
-method(contents_markdown, new_S3_class("Chat")) <- function(content, heading_level = 2) {
+method(contents_markdown, new_S3_class("Chat")) <- function(
+  content,
+  heading_level = 2
+) {
   turns <- content$get_turns()
   if (length(turns) == 0) {
     return("")
@@ -604,7 +698,7 @@ method(contents_markdown, new_S3_class("Chat")) <- function(content, heading_lev
     res[i] <- glue::glue("{hh} {role}\n\n{contents_markdown(turns[[i]])}")
   }
 
-  paste(res, collapse="\n\n")
+  paste(res, collapse = "\n\n")
 }
 
 extract_data <- function(turn, type, convert = TRUE, needs_wrapper = FALSE) {