+++ R/provider.R
@@ -14,6 +14,7 @@ NULL
 #'
 #' @export
 #' @param base_url The base URL for the API.
+#' @param params A list of standard parameters created by [params()].
 #' @param extra_args Arbitrary extra arguments to be included in the request body.
 #' @return An S7 Provider object.
 #' @examples
@@ -22,19 +23,30 @@ Provider <- new_class(
   "Provider",
   properties = list(
     base_url = prop_string(),
+    params = class_list,
     extra_args = class_list
   )
 )
 
 # Create a request------------------------------------
 
-chat_request <- new_generic("chat_request", "provider",
-  function(provider, stream = TRUE, turns = list(), tools = list(), type = NULL) {
+chat_request <- new_generic(
+  "chat_request",
+  "provider",
+  function(
+    provider,
+    stream = TRUE,
+    turns = list(),
+    tools = list(),
+    type = NULL
+  ) {
     S7_dispatch()
   }
 )
 
-chat_resp_stream <- new_generic("chat_resp_stream", "provider",
+chat_resp_stream <- new_generic(
+  "chat_resp_stream",
+  "provider",
   function(provider, resp) {
     S7_dispatch()
   }
@@ -43,19 +55,33 @@ method(chat_resp_stream, Provider) <- function(provider, resp) {
   resp_stream_sse(resp)
 }
 
+chat_params <- new_generic(
+  "chat_params",
+  "provider",
+  function(provider, params) {
+    S7_dispatch()
+  }
+)
+
 # Extract data from streaming results ------------------------------------
 
-stream_parse <- new_generic("stream_parse", "provider",
+stream_parse <- new_generic(
+  "stream_parse",
+  "provider",
   function(provider, event) {
     S7_dispatch()
   }
 )
-stream_text <- new_generic("stream_text", "provider",
+stream_text <- new_generic(
+  "stream_text",
+  "provider",
   function(provider, event) {
     S7_dispatch()
   }
 )
-stream_merge_chunks <- new_generic("stream_merge_chunks", "provider",
+stream_merge_chunks <- new_generic(
+  "stream_merge_chunks",
+  "provider",
   function(provider, result, chunk) {
     S7_dispatch()
   }