+++ tests/testthat/helper-provider.R
@@ -17,6 +17,14 @@ retry_test <- function(code, retries = 1) {
   eval(get_expr(code), get_env(code))
 }
 
+# Params -----------------------------------------------------------------
+
+test_params_stop <- function(chat_fun) {
+  chat <- chat_fun(params = params(stop_sequences = "cool"))
+  out <- chat$chat("Repeat after the following phrase: Dogs are cool")
+  expect_equal(out, "Dogs are ")
+}
+
 # Turns ------------------------------------------------------------------
 
 test_turns_system <- function(chat_fun) {
@@ -34,11 +42,16 @@ test_turns_system <- function(chat_fun) {
 }
 
 test_turns_existing <- function(chat_fun) {
-  chat <- chat_fun(turns = list(
-    Turn("system", "Return very minimal output; no punctuation."),
-    Turn("user", "List the names of any 8 of Santa's 9 reindeer."),
-    Turn("assistant", "Dasher, Dancer, Vixen, Comet, Cupid, Donner, Blitzen, and Rudolph.")
-  ))
+  chat <- chat_fun(
+    turns = list(
+      Turn("system", "Return very minimal output; no punctuation."),
+      Turn("user", "List the names of any 8 of Santa's 9 reindeer."),
+      Turn(
+        "assistant",
+        "Dasher, Dancer, Vixen, Comet, Cupid, Donner, Blitzen, and Rudolph."
+      )
+    )
+  )
   expect_length(chat$get_turns(), 2)
 
   resp <- chat$chat("Who is the remaining one? Just give the name")
@@ -64,7 +77,10 @@ test_tools_simple <- function(chat_fun) {
 
 test_tools_async <- function(chat_fun) {
   chat <- chat_fun(system_prompt = "Be very terse, not even punctuation.")
-  chat$register_tool(tool(coro::async(function() "2024-01-01"), "Return the current date"))
+  chat$register_tool(tool(
+    coro::async(function() "2024-01-01"),
+    "Return the current date"
+  ))
 
   result <- sync(chat$chat_async("What's the current date in YMD format?"))
   expect_match(result, "2024-01-01")
@@ -83,25 +99,30 @@ test_tools_parallel <- function(chat_fun) {
     person = type_string("Name of a person")
   ))
 
-  result <- chat$chat("
+  result <- chat$chat(
+    "
     What are Joe and Hadley's favourite colours?
     Answer like name1: colour1, name2: colour2
-  ")
+  "
+  )
   expect_match(result, "Joe: sage green")
   expect_match(result, "Hadley: red")
   expect_length(chat$get_turns(), 4)
 }
 
 test_tools_sequential <- function(chat_fun, total_calls) {
-  chat <- chat_fun(system_prompt = "
+  chat <- chat_fun(
+    system_prompt = "
     Use provided tool calls to find the weather forecast and suitable
     equipment for a variety of weather conditions.
 
     In your response, be very terse and omit punctuation.
-  ")
+  "
+  )
 
   forecast <- function(city) if (city == "New York") "rainy" else "sunny"
-  equipment <- function(weather) if (weather == "rainy") "umbrella" else "sunscreen"
+  equipment <- function(weather)
+    if (weather == "rainy") "umbrella" else "sunscreen"
   chat$register_tool(tool(
     forecast,
     "Gets the weather forecast for a city",
@@ -137,11 +158,17 @@ test_data_extraction <- function(chat_fun) {
 
   chat <- chat_fun()
   data <- chat$extract_data(prompt, type = article_summary)
-  expect_mapequal(data, list(title = "Apples are tasty", author = "Hadley Wickham"))
+  expect_mapequal(
+    data,
+    list(title = "Apples are tasty", author = "Hadley Wickham")
+  )
 
   # Check that we can do it again
   data <- chat$extract_data(prompt, type = article_summary)
-  expect_mapequal(data, list(title = "Apples are tasty", author = "Hadley Wickham"))
+  expect_mapequal(
+    data,
+    list(title = "Apples are tasty", author = "Hadley Wickham")
+  )
 }
 
 # Images -----------------------------------------------------------------