+++ tests/testthat/test-chat.R
@@ -1,8 +1,10 @@
 test_that("can get and set the system prompt", {
-  chat <- chat_openai(turns = list(
-    Turn("user", "Hi"),
-    Turn("assistant", "Hello")
-  ))
+  chat <- chat_openai_test(
+    turns = list(
+      Turn("user", "Hi"),
+      Turn("assistant", "Hello")
+    )
+  )
 
   # NULL -> NULL
   chat$set_system_prompt(NULL)
@@ -22,7 +24,7 @@ test_that("can get and set the system prompt", {
 })
 
 test_that("can retrieve system prompt with last_turn()", {
-  chat1 <- chat_openai()
+  chat1 <- chat_openai_test()
   expect_equal(chat1$last_turn("system"), NULL)
 
   chat2 <- chat_openai(system_prompt = "You are from New Zealand")
@@ -46,12 +48,12 @@ test_that("can get and set turns", {
 })
 
 test_that("can get model", {
-  chat <- chat_openai(model = "abc")
+  chat <- chat_openai_test(model = "abc")
   expect_equal(chat$get_model(), "abc")
 })
 
 test_that("setting turns usually preserves, but can set system prompt", {
-  chat <- chat_openai(system_prompt = "You're a funny guy")
+  chat <- chat_openai_test(system_prompt = "You're a funny guy")
   chat$set_turns(list())
   expect_equal(chat$get_system_prompt(), "You're a funny guy")
 
@@ -60,7 +62,7 @@ test_that("setting turns usually preserves, but can set system prompt", {
 })
 
 test_that("can perform a simple batch chat", {
-  chat <- chat_openai()
+  chat <- chat_openai_test()
 
   result <- chat$chat("What's 1 + 1. Just give me the answer, no punctuation")
   expect_equal(result, "2")
@@ -68,9 +70,11 @@ test_that("can perform a simple batch chat", {
 })
 
 test_that("can perform a simple async batch chat", {
-  chat <- chat_openai()
+  chat <- chat_openai_test()
 
-  result <- chat$chat_async("What's 1 + 1. Just give me the answer, no punctuation")
+  result <- chat$chat_async(
+    "What's 1 + 1. Just give me the answer, no punctuation"
+  )
   expect_s3_class(result, "promise")
 
   result <- sync(result)
@@ -79,37 +83,49 @@ test_that("can perform a simple async batch chat", {
 })
 
 test_that("can perform a simple streaming chat", {
-  chat <- chat_openai()
+  chat <- chat_openai_test()
 
-  chunks <- coro::collect(chat$stream("
+  chunks <- coro::collect(chat$stream(
+    "
     What are the canonical colors of the ROYGBIV rainbow?
     Put each colour on its own line. Don't use punctuation.
-  "))
+  "
+  ))
   expect_gt(length(chunks), 2)
 
   rainbow_re <- "^red *\norange *\nyellow *\ngreen *\nblue *\nindigo *\nviolet *\n?$"
   expect_match(paste(chunks, collapse = ""), rainbow_re, ignore.case = TRUE)
-  expect_match(chat$last_turn()@contents[[1]]@text, rainbow_re, ignore.case = TRUE)
+  expect_match(
+    chat$last_turn()@contents[[1]]@text,
+    rainbow_re,
+    ignore.case = TRUE
+  )
 })
 
 test_that("can perform a simple async batch chat", {
-  chat <- chat_openai()
+  chat <- chat_openai_test()
 
-  chunks <- coro::async_collect(chat$stream_async("
+  chunks <- coro::async_collect(chat$stream_async(
+    "
     What are the canonical colors of the ROYGBIV rainbow?
     Put each colour on its own line. Don't use punctuation.
-  "))
+  "
+  ))
   expect_s3_class(chunks, "promise")
 
   chunks <- sync(chunks)
   expect_gt(length(chunks), 2)
   rainbow_re <- "^red *\norange *\nyellow *\ngreen *\nblue *\nindigo *\nviolet *\n?$"
   expect_match(paste(chunks, collapse = ""), rainbow_re, ignore.case = TRUE)
-  expect_match(chat$last_turn()@contents[[1]]@text, rainbow_re, ignore.case = TRUE)
+  expect_match(
+    chat$last_turn()@contents[[1]]@text,
+    rainbow_re,
+    ignore.case = TRUE
+  )
 })
 
 test_that("can chat in parallel", {
-  chat <- chat_openai("Just give me answers, no punctuation")
+  chat <- chat_openai_test("Just give me answers, no punctuation")
   results <- chat$chat_parallel(list("What's 1 + 1?", "What's 2 + 2?"))
 
   expect_type(results, "list")
@@ -125,7 +141,7 @@ test_that("can chat in parallel", {
 test_that("can extract structured data", {
   person <- type_object(name = type_string(), age = type_integer())
 
-  chat <- chat_openai()
+  chat <- chat_openai_test()
   data <- chat$extract_data("John, age 15, won first prize", type = person)
   expect_equal(data, list(name = "John", age = 15))
 })
@@ -133,7 +149,7 @@ test_that("can extract structured data", {
 test_that("can extract data in parallel", {
   person <- type_object(name = type_string(), age = type_integer())
 
-  chat <- chat_openai()
+  chat <- chat_openai_test()
   data <- chat$extract_data_parallel(
     list(
       "John, age 15, won first prize",
@@ -153,9 +169,22 @@ test_that("can extract data in parallel", {
 test_that("can extract structured data (async)", {
   person <- type_object(name = type_string(), age = type_integer())
 
-  chat <- chat_openai()
-  data <- sync(chat$extract_data_async("John, age 15, won first prize", type = person))
+  chat <- chat_openai_test()
+  data <- sync(chat$extract_data_async(
+    "John, age 15, won first prize",
+    type = person
+  ))
   expect_equal(data, list(name = "John", age = 15))
+})
+
+test_that("can retrieve tokens with or without system prompt", {
+  chat <- chat_openai_test("abc")
+  expect_equal(nrow(chat$tokens(FALSE)), 0)
+  expect_equal(nrow(chat$tokens(TRUE)), 1)
+
+  chat <- chat_openai()
+  expect_equal(nrow(chat$tokens(FALSE)), 0)
+  expect_equal(nrow(chat$tokens(TRUE)), 0)
 })
 
 test_that("has a basic print method", {
@@ -236,4 +265,53 @@ test_that("async chat messages get timestamped in sequence", {
   expect_true(turns[[1]]@completed >= before_send)
   expect_true(turns[[1]]@completed <= turns[[2]]@completed)
   expect_true(turns[[2]]@completed <= after_receive)
+})
+
+test_that("chat can get and register a list of tools", {
+  chat <- chat_openai(api_key = "not required")
+  chat2 <- chat_openai(api_key = "not required")
+
+  tools <- list(
+    "sys_time" = tool(
+      function() strftime(Sys.time(), "%F %T"),
+      .description = "Get the current system time",
+      .name = "sys_time"
+    ),
+    "r_version" = tool(
+      function() R.version.string,
+      .description = "Get the R version of the current session",
+      .name = "r_version"
+    )
+  )
+
+  for (tool in tools) {
+    chat$register_tool(tool)
+  }
+
+  chat2$set_tools(tools)
+
+  expect_equal(chat$get_tools(), tools)
+  expect_equal(chat2$get_tools(), chat$get_tools())
+
+  # action = "replace" overwrites existing tools
+  tool_r_major <- tool(
+    function() R.version$major,
+    .description = "Get the major version of R",
+    .name = "r_version_major"
+  )
+  new_tools <- list("r_version_major" = tool_r_major)
+  chat$set_tools(new_tools)
+  expect_equal(chat$get_tools(), new_tools)
+
+  # set_tools() throws with helpful message if given just a tool
+  expect_snapshot(
+    error = TRUE,
+    chat$set_tools(tools[[1]])
+  )
+
+  # set_tools() throws with helpful message if not all items are tools
+  expect_snapshot(
+    error = TRUE,
+    chat$set_tools(c(tools, list("foo")))
+  )
 })