+++ tests/testthat/test-provider-gemini.R
@@ -19,6 +19,12 @@ test_that("defaults are reported", {
   expect_snapshot(. <- chat_gemini())
 })
 
+test_that("supports standard parameters", {
+  chat_fun <- chat_gemini
+
+  test_params_stop(chat_fun)
+})
+
 test_that("respects turns interface", {
   chat_fun <- chat_gemini
 
@@ -69,10 +75,13 @@ test_that("can merge text output", {
   out <- merge_gemini_chunks(out, chunks[[3]])
 
   expect_equal(out$candidates[[1]]$content$parts[[1]]$text, "abc")
-  expect_equal(out$usageMetadata, list(
-    promptTokenCount = 5,
-    candidatesTokenCount = 17,
-    totalTokenCount = 22
-  ))
+  expect_equal(
+    out$usageMetadata,
+    list(
+      promptTokenCount = 5,
+      candidatesTokenCount = 17,
+      totalTokenCount = 22
+    )
+  )
   expect_equal(out$candidates[[1]]$finishReason, "STOP")
 })