+++ tests/testthat/test-types.R
@@ -17,13 +17,22 @@ test_that("can convert arrays of basic types to simple vectors", {
   )
 })
 
+test_that("can convert arrays of optional basic types to simple vectors", {
+  type <- type_array(items = type_boolean(required = FALSE))
+  expect_equal(convert_from_type(list(FALSE, TRUE), type), c(FALSE, TRUE))
+  expect_equal(convert_from_type(list(NULL, TRUE), type), c(NA, TRUE))
+
+  type <- type_array(items = type_integer(required = FALSE))
+  expect_identical(convert_from_type(list(NULL), type), NA_integer_)
+})
+
 test_that("can covert array of arrays to lists of vectors", {
   expect_equal(
     convert_from_type(
-       list(list(1, 2), list(3, 4)),
-       type_array(items = type_array(items = type_integer()))
-     ),
-     list(c(1L, 2L), c(3L, 4L))
+      list(list(1, 2), list(3, 4)),
+      type_array(items = type_array(items = type_integer()))
+    ),
+    list(c(1L, 2L), c(3L, 4L))
   )
 })
 
@@ -41,10 +50,12 @@ test_that("can covert arrays of objects to data framnes", {
   expect_equal(
     convert_from_type(
       list(list(x = 1, y = "x"), list(x = 3, y = "y")),
-      type_array(items = type_object(
-        x = type_integer(),
-        y = type_string()
-      ))
+      type_array(
+        items = type_object(
+          x = type_integer(),
+          y = type_string()
+        )
+      )
     ),
     data.frame(x = c(1L, 3L), y = c("x", "y"))
   )